[Impeller] Remove libtess2 from libflutter.
