+++ impeller/tessellator/BUILD.gn
@@ -15,6 +15,21 @@ impeller_component("tessellator") {
   deps = [
     "../core",
     "//flutter/fml",
+  ]
+}
+
+impeller_component("tessellator_libtess") {
+  sources = [
+    "tessellator_libtess.cc",
+    "tessellator_libtess.h",
+  ]
+
+  public_deps = [ "../geometry" ]
+
+  deps = [
+    ":tessellator",
+    "../core",
+    "//flutter/fml",
     "//third_party/libtess2",
   ]
 }
@@ -30,11 +45,11 @@ impeller_component("tessellator_shared") {
   sources = [
     "c/tessellator.cc",
     "c/tessellator.h",
-    "tessellator.cc",
-    "tessellator.h",
   ]
 
   deps = [
+    ":tessellator",
+    ":tessellator_libtess",
     "../core",
     "../geometry",
     "//flutter/fml",
@@ -50,7 +65,7 @@ impeller_component("tessellator_unittests") {
   testonly = true
   sources = [ "tessellator_unittests.cc" ]
   deps = [
-    ":tessellator",
+    ":tessellator_libtess",
     "../geometry:geometry_asserts",
     "//flutter/testing",
   ]
